[config_mgmt.py]: Shutdown the interfaces after config validation.
Changes:
-- Shutdown the interfaces after config validation while Dy Port Breakout.
-- lib.py: to avoid loop dependencies b/w config_mgmt.py and main.py
   moved few helper funtions in lib.py.
