+++ config/config_mgmt.py
@@ -10,6 +10,8 @@ try:
     load_source('sonic_cfggen', '/usr/local/bin/sonic-cfggen')
     from sonic_cfggen import deep_update, FormatConverter, sort_data
     from swsssdk import ConfigDBConnector, SonicV2Connector, port_util
+    from lib import interface_name_is_valid
+
     from pprint import PrettyPrinter, pprint
     from json import dump, load, dumps, loads
     from sys import path as sysPath
@@ -232,9 +234,11 @@ class configMgmt():
             if_name_map, if_oid_map = port_util.get_interface_oid_map(dataBase)
             self.sysLog(syslog.LOG_DEBUG, 'if_name_map {}'.format(if_name_map))
 
-            # If we are here, then get ready to update the Config DB, Update
-            # deletion of Config first, then verify in Asic DB for port deletion,
+            # If we are here, then get ready to update the Config DB,
+            # shutdown the ports, Update deletion of Config,
+            # verify in Asic DB for port deletion,
             # then update addition of ports in config DB.
+            self.shutdownIntf(delPorts)
             self.writeConfigDB(delConfigToLoad)
             # Verify in Asic DB,
             self.verifyAsicDB(db=dataBase, ports=delPorts, portMap=if_name_map, \
@@ -384,6 +388,23 @@ class configMgmt():
         print('Data Validation successful')
         self.sysLog(msg='Data Validation successful')
         return True
+
+    """
+    Based on the list of Ports, create a dict to shutdown port, update Config DB.
+    Input: [] of ports.
+    """
+    def shutdownIntf(self, ports):
+        """ shut down all the interfaces before deletion """
+        shutDownConf = dict(); shutDownConf["PORT"] = dict()
+        for intf in ports:
+            if not interface_name_is_valid(intf):
+                raise Exception("Interface name is invalid")
+            shutDownConf["PORT"][intf] = {"admin_status": "down"}
+        if len(shutDownConf["PORT"]):
+            self.sysLog(msg='shutdown Interfaces: {}'.format(shutDownConf))
+            self.writeConfigDB(shutDownConf)
+
+        return
 
     """
     Merge second dict in first, Note both first and second dict will be changed
